Fix an issue with configuration delegation in Rails < 8.1
A bug was introduced in 2cf722bd7cc310009ee626cce28a7ea52b47f649 that
caused `TokenVerifier` to not work correctly in Rails < 8.1.

After upgrading to the latest version of this gem, users would see
`ActionController::InvalidAuthenticityToken` error during the request
phrase, causing the OAuth authentication flow to fail.

The root cause of this issue was that the line
`include ActionController::RequestForgeryProtection` got moved below
the code that sets up the delegation to `ActionController::Base.config`.
This caused the delegation setup to silently fail as the list of
configurations to set up the delegation became empty.

Bug investigation: https://github.com/cookpad/omniauth-rails_csrf_protection/issues/26#issuecomment-3591652174

This issue was not caught by the test suite because the default
configuration values in `ActionController::RequestForgeryProtection`
and `ActionController::Base.config` were the same in the test suite.
However, in an actual Ruby on Rails application, Railtie would set
`per_form_csrf_tokens` to `true` on Action Controller during
application's initialization[1], causing the configuration to go
out-of-sync.

[1]: https://github.com/rails/rails/blob/v8.0.0/railties/lib/rails/application/configuration.rb#L114

This commit fixes the issue by moving `include` to the right location
for Rails 8.1+ and Rails < 8.1. To prevent further regression, this
commit also includes a new integration test that renders a mock sign-in
form to test the whole sign-in flow.

I'd like to thank @trysmr for a detailed investigation on the issue,
which helped me understand the root cause and also found the reason why
our test case didn't catch it.

Fixes #26.